Update README.md
Result

The analysis of Bitcoin's closing price trend over the course of 5 years has revealed significant fluctuations. On March 11, 2020, the highest price increase of 59% was recorded, while on February 7, 2021, the lowest decrease of -15% occurred.

The probable causes behind the highest price increase on March 11, 2020, are likely attributed to several factors, including increased interest and investment in cryptocurrencies, advancements in blockchain technology, and potentially the impact of global events such as the COVID-19 pandemic, which spurred investor interest in digital assets as an alternative investment.

On the other hand, the largest price decrease on February 7, 2021, could have been caused by various factors, including a correction in price following a significant prior uptrend, negative news or events in the crypto space, or massive selling actions by investors. It's important to note that the crypto market is known for its high volatility, making significant price fluctuations like these not uncommon.